Refactor ListTrackedItems by dividing into two
